requirements.txt
setuptools~=50.3.2
requests~=2.22.0
Keras~=2.2.5
numpy~=1.19.4
nltk~=3.5
six~=1.15.0
gensim~=3.8.3
scikit-learn~=0.22.0
nose~=1.3.7
scipy~=1.3.3
Click~=7.0
tqdm~=4.51.0
regex~=2020.11.11
pyparsing~=2.4.6
Unidecode~=1.1.1
packaging~=20.3
PyGithub~=1.54.1
multidict~=4.7.6
yarl~=1.4.2
attrs~=19.3.0
gunicorn~=20.0.4
aiodns~=2.0.0
cchardet~=2.1.6
pytest~=5.4.2